requirements.txt
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
python-prctl==1.7
pytz==2020.1
six==1.15.0
